rename nav folder to PlanTrip
