+++ src/Components/App/App.js
@@ -5,7 +5,7 @@ import Register from '../Register';
 import Header from '../Header';
 import Dashboard from '../Dashboard';
 import NotFound from '../NotFound';
-import PlanTrip from '../Nav/PlanTrip';
+import PlanTrip from '../PlanTrip/PlanTrip';
 import LandingPage from '../LandingRoute/LandingPage';
 import MapContainer from '../MapContainer/MapContainer';
 import Interests from '../../Interests/Interests';
@@ -15,10 +15,6 @@ import './App.css';
 import MyTrips from "../myTrips/MyTrips"
 
 class App extends React.Component {
-  // constructor(props) {
-  //   super(props);
-
-  // }
 
   render() {
